requirements.txt
absl-py==0.8.1
astor==0.8.0
astroid==2.3.3
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
imageio==2.5.0
imutils==0.5.2
isort==4.3.21
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
matplotlib==3.0.3
mccabe==0.6.1
networkx==2.3
numpy==1.16.2
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
opt-einsum==3.1.0
Pillow==6.0.0
protobuf==3.10.0
pylint==2.4.3
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
selectivesearch==0.4
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
typed-ast==1.4.0
Werkzeug==0.16.0
wrapt==1.11.2
